A little temporary regorg and eeprom wear leveing kinda
Reorg as notepad++ and arduinoide no likey work together, so hiding the
files from arduino ide for now.

SyrupSettingsManager to grab settings and do fun settingsy things.

+++ Arduino/libraries/Sensors/TempProbe.h
@@ -23,8 +23,8 @@ class TempProbe : public Subject<TempProbe>
       SCALE,
       TEMP
     };
-    double m_tempF;
-    double m_tempC;
+    float m_tempF;
+    float m_tempC;
     int m_tempProbeReading;
     int m_tempProbePin;
     int m_digitalMappedTempProbePin;
@@ -38,9 +38,9 @@ class TempProbe : public Subject<TempProbe>
     void pause();
     void unpause();
     void tick();
-    double convertReadingToF(int tempProbeReading);
-    double convertReadingToF(double tempC);
-    double convertReadingToC(int tempProbeReading);
+    float convertReadingToF(int tempProbeReading);
+    float convertReadingToF(float tempC);
+    float convertReadingToC(int tempProbeReading);
     
   private:
     e_scale m_activeScale;
